requirements.txt
python-dotenv
gradio
torch
torchvision
Pillow
numpy